refactor: isValueNaN関数をMathからValidateに移動
