+++ package/main/src/Validate/isValueNaN.ts
